+++ src/aspects.lua
@@ -455,7 +455,7 @@ mod.AspectTraitData = {
 		},
 		AddOutgoingDamageModifiers =
 		{
-			ValidWeapons = { "WeaponAxeSpin", "WeaponStaffSwing5", "WeaponDagger5"},
+			ValidWeapons = { "WeaponAxeSpin", "WeaponStaffSwing5", "WeaponDagger5", "WeaponTorch", "WeaponLob"},
 			ValidSuitProjectile = true,
 		},
 		OnProjectileDeathFunction =
@@ -565,7 +565,7 @@ mod.AspectTraitData = {
 		},
 		OnWeaponFiredFunctions =
 		{
-			ValidWeapons = {"WeaponSuitCharged", "WeaponSuitRanged", "WeaponAxeSpin", "WeaponStaffSwing5", "WeaponDagger5" },
+			ValidWeapons = {"WeaponSuitCharged", "WeaponSuitRanged", "WeaponAxeSpin", "WeaponStaffSwing5", "WeaponDagger5", "WeaponLob", "WeaponTorch" },
 			FunctionName = "CheckSuitComboAttackBuff",
 			FunctionArgs =
 			{
@@ -581,7 +581,10 @@ mod.AspectTraitData = {
 		},
 		OnProjectileCreationFunction =
 		{
-			ValidProjectiles = { "ProjectileSwing5", "ProjectileAxeSpin" },
+			ValidProjectiles = {
+				"ProjectileSwing5", "ProjectileAxeSpin", "ProjectileTorchWave", "ProjectileTorchGhostLarge", "ProjectileTorchSupayBallEx", "ProjectileTorchBallEos",
+				"ProjectileLobCharged", "ProjectileLobOverheat"
+			},
 			Name = _PLUGIN.guid .. "." .. "CheckSuitComboAttackBuff",
 			Args =
 			{
@@ -2183,6 +2186,182 @@ mod.AspectTraitData = {
 			},
 		},
 		FlavorText = "DaggerBlockAspect_FlavorText",
+	},
+
+	LobImpulseAspect_Secondary =
+	{
+		InheritFrom = {"BaseTrait"},
+		PreEquipWeapons = {"WeaponSkullImpulse"},
+		Icon = "Hammer_Lob_15",
+		ReplacementGrannyModels = 
+		{
+			WeaponLob_Mesh = "WeaponLob_Persephone_Mesh"
+		},
+		Charge = 0,
+		PropertyChanges = {
+			{
+				WeaponName = "WeaponLobSpecial",
+				ProjectileName = "ProjectileThrowCharged",
+				ProjectileProperty = "Graphic",
+				ChangeValue = "LobSpecialFx_Persephone",
+				ChangeType = "Absolute",
+			},
+			{
+				WeaponName = "WeaponLobSpecial",
+				ProjectileName = "ProjectileThrowBlink",
+				ProjectileProperty = "Graphic",
+				ChangeValue = "DashLobTrailEmitter_Persephone",
+				ChangeType = "Absolute",
+			},
+			{
+				WeaponName = "WeaponSkullImpulse",
+				ProjectileName = "ProjectileSkullImpulse",
+				ProjectileProperty = "Graphic",
+				ChangeValue = "DashLobTrailEmitter_Persephone",
+				ChangeType = "Absolute",
+			},
+		},
+		RarityLevels =
+		{
+			Common =
+			{
+				Multiplier = 1,
+			},
+			Rare =
+			{
+				Multiplier = 1.5,
+			},
+			Epic =
+			{
+				Multiplier = 2,
+			},
+			Heroic =
+			{
+				Multiplier = 5/2,
+			},
+			Legendary =
+			{
+				Multiplier = 6/2,
+			},
+			Perfect =
+			{
+				Multiplier = 9/2,
+			},
+		},
+		SetupFunction =
+		{
+			Threaded = true,
+			Name = "SetupSkullImpulseUI",
+		},
+		OnWeaponFiredFunctions = 
+		{
+			ValidWeapons = {"WeaponLobSpecial", "WeaponSkullImpulse"},
+			FunctionName = "SkullImpulseTransform",
+			FunctionArgs = 
+			{
+				BaseDuration = 0.5,			-- Duration of ex attack w/ no charge
+				Interval = 200,
+			}
+		},
+		MaxBonusBoonRankWeighted =
+		{
+			BaseValue = 2,
+		},
+		MaxBonusBoonRankDistribution =
+		{
+			[2] =
+			{
+				-- Weighted list, so all values should add up to 1 for best distribution
+				-- Don't add entries for 1 because that's the same as 0 boon and causes 'level 1' to show up on boon menus
+				[0] = 0.70,
+				[2] = 0.30,
+			},
+			[3] =
+			{
+				[0] = 0.65,
+				[2] = 0.30,
+				[3] = 0.05,
+			},
+			[4] =
+			{
+				[0] = 0.60,
+				[2] = 0.25,
+				[3] = 0.10,
+				[4] = 0.05,
+			},
+			[5] =
+			{
+				[0] = 0.55,
+				[2] = 0.20,
+				[3] = 0.15,
+				[4] = 0.10,
+				[5] = 0.05,
+			},
+			[6] =
+			{
+				[0] = 0.50,
+				[2] = 0.16,
+				[3] = 0.14,
+				[4] = 0.12,
+				[5] = 0.06,
+				[6] = 0.02,
+			},
+			[9] =
+			{
+				[0] = 0.28,
+				[2] = 0.16,
+				[3] = 0.14,
+				[4] = 0.12,
+				[5] = 0.10,
+				[6] = 0.08,
+				[7] = 0.06,
+				[8] = 0.04,
+				[9] = 0.02,
+			}
+		},
+		OnEnemyDamagedAction = 
+		{
+			AllEffectsTrigger = true,
+			FunctionName = "ChargeSkullImpulse",
+			Args = 
+			{
+				-- One "charge" is equal to full one second of skull car
+				-- 0.001 means 100 damage = 0.1 seconds of skull car charge
+				ValidProjectiles = game.WeaponSets.OlympianProjectileNames,
+				ValidEffectNames = game.WeaponSets.OlympianEffectNames,
+				ChargePerDamage = 0.00100,
+				MaxCharge = 2,
+				MinChargeToFire = 0.5,
+				ReportValues =
+				{
+					ReportedMinChargeToFire = "MinChargeToFire",
+					ReportedMaxCharge = "MaxCharge",
+				}
+			}
+		},
+		StatLines =
+		{
+			"ExDamageStatDisplay1",
+		},
+		ExtractValues =
+		{
+			{
+				Key = "MaxBonusBoonRankWeighted",
+				ExtractAs = "BoonRank",
+				IncludeSigns = true,
+			},
+			{
+				Key = "ReportedMaxCharge",
+				ExtractAs = "Duration",
+				SkipAutoExtract = true,
+			},
+			{
+				Key = "ReportedMinChargeToFire",
+				ExtractAs = "MinCharge",
+				SkipAutoExtract = true,
+			},
+		},
+		FlavorText = "LobImpulseAspect_FlavorText",
 	}
 }
 
@@ -2383,7 +2562,8 @@ modutil.mod.Path.Wrap("ShivaAttackBoostClear", function (base, triggerArgs)
 end)
 
 modutil.mod.Path.Wrap("CheckSuitComboAttackBuff", function (base, weaponData, functionArgs, triggerArgs)
-	if game.Contains({"WeaponStaffSwing5", "WeaponAxeSpin"}, weaponData.Name) then
+	if game.Contains({"WeaponAxeSpin", "WeaponStaffSwing5", "WeaponLob", "WeaponTorch"}, weaponData.Name) and
+			game.IsExWeapon( weaponData.Name, { Combat = true }, triggerArgs ) then
 		if weaponData.Name == "WeaponStaffSwing5" then
 			game.waitUntil(_PLUGIN.guid .. "ProjectileCreation")
 			game.waitUntil(_PLUGIN.guid .. "ProjectileCreation")
@@ -2393,11 +2573,16 @@ modutil.mod.Path.Wrap("CheckSuitComboAttackBuff", function (base, weaponData, fu
 			end
 		elseif weaponData.Name == "WeaponAxeSpin" then
 			game.waitUntil(_PLUGIN.guid .. "ProjectileCreation")
+		elseif weaponData.Name == "WeaponTorch" then
+			game.wait(0.01)
+		elseif weaponData.Name == "WeaponLob" then
+
 		end
 
 		local projectileIds = game.SessionMapState[_PLUGIN.guid .. "ProjectileIds"] or {}
 		local stacks = game.CurrentRun.Hero.ActiveEffects[functionArgs.EffectName]
 		if not stacks then
+			game.SessionMapState[_PLUGIN.guid .. "ProjectileIds"] = nil
 			return
 		end
 		if functionArgs.SelfEffectMaxStacks and stacks > functionArgs.SelfEffectMaxStacks then
@@ -2407,6 +2592,9 @@ modutil.mod.Path.Wrap("CheckSuitComboAttackBuff", function (base, weaponData, fu
 			game.SessionMapState.SuitBonusProjectileId[id] = 1 + functionArgs.SelfEffectStackMultiplier * stacks
 		end
 		game.ClearEffect({Id = game.CurrentRun.Hero.ObjectId, Name = "ShivaAttackBoost"})
+		game.SessionMapState[_PLUGIN.guid .. "ProjectileIds"] = nil
+		return
+	elseif game.Contains({"WeaponAxeSpin", "WeaponStaffSwing5", "WeaponLob", "WeaponTorch"}, weaponData.Name) then
 		game.SessionMapState[_PLUGIN.guid .. "ProjectileIds"] = nil
 		return
 	end
@@ -2467,6 +2655,37 @@ modutil.mod.Path.Wrap("ShowDaggerUI", function (base, args)
 	game.SetAlpha({ Id = game.ScreenAnchors.DaggerUI, Duration = game.HUDScreen.FadeInDuration, Fraction = game.ConfigOptionCache.HUDOpacity })
 	game.SetAlpha({ Id = game.ScreenAnchors.DaggerUIChargeAmount, Duration = 0, Fraction = 0 })
 	game.SetAlpha({ Id = game.ScreenAnchors.DaggerUIChargeAmount, Duration = game.HUDScreen.FadeInDuration, Fraction = game.ConfigOptionCache.HUDOpacity })
+end)
+
+modutil.mod.Path.Wrap("ShowLobUI", function (base)
+	base()
+	if (not game.HeroHasTrait("LobImpulseAspect_Secondary") and not game.HeroHasTrait("LobGunAspect_Secondary")) or not game.ShowingCombatUI then
+		return
+	end
+
+	if game.ScreenAnchors.LobUI ~= nil then
+		return
+	end
+	game.ScreenAnchors.LobUI = game.CreateScreenObstacle({ Name = "BlankObstacle", Group = "Combat_Menu_TraitTray_Overlay_Additive", X = game.HUDScreen.AmmoX + 150, Y = game.ScreenHeight - game.HUDScreen.AmmoBottomOffset })
+	game.ScreenAnchors.LobUIChargeAmount = game.CreateScreenObstacle({ Name = "BlankObstacle", Group = game.HUDScreen.ComponentData.DefaultGroup, X = game.HUDScreen.AmmoX + 150, Y = game.ScreenHeight - game.HUDScreen.AmmoBottomOffset })
+	game.SetAnimation({ Name = "StaffReloadTimer", DestinationId = game.ScreenAnchors.LobUIChargeAmount })
+	if game.HeroHasTrait("LobImpulseAspect_Secondary") then
+		local trait = game.GetHeroTrait("LobImpulseAspect_Secondary")
+		local currentCharge = trait.Charge
+		local maxCharge = trait.OnEnemyDamagedAction.Args.MaxCharge
+		local currentChargeText = game.round( currentCharge, 1)
+		game.SetAnimationFrameTarget({ Name = "StaffReloadTimer", DestinationId = game.ScreenAnchors.LobUIChargeAmount, Fraction = currentCharge/ maxCharge, Instant = true })
+
+		if trait.Charge >= trait.OnEnemyDamagedAction.Args.MaxCharge then
+			game.SetAnimation({ Name = "StaffReloadTimerReady", SuppressSounds = true, DestinationId = game.ScreenAnchors.LobUI })
+		end
+	elseif game.HeroHasTrait("LobGunAspect_Secondary") then
+		game.SetAnimationFrameTarget({ Name = "StaffReloadTimer", DestinationId = game.ScreenAnchors.LobUIChargeAmount, Fraction = 0, Instant = true })
+	end
+	game.SetAlpha({ Id = game.ScreenAnchors.LobUI, Duration = 0, Fraction = 0 })
+	game.SetAlpha({ Id = game.ScreenAnchors.LobUI, Duration = game.HUDScreen.FadeInDuration, Fraction = game.ConfigOptionCache.HUDOpacity })
+	game.SetAlpha({ Id = game.ScreenAnchors.LobUIChargeAmount, Duration = 0, Fraction = 0 })
+	game.SetAlpha({ Id = game.ScreenAnchors.LobUIChargeAmount, Duration = game.HUDScreen.FadeInDuration, Fraction = game.ConfigOptionCache.HUDOpacity })
 end)
 
 game.OnWeaponChargeCanceled{ "WeaponAxeSpin",
@@ -2540,3 +2759,50 @@ game.OnWeaponChargeCanceled{ "WeaponSuitCharged",
 		end
 	end
 }
+
+game.TraitData.LobCloseAttackAspect.OnEnemyDamagedAction.ValidWeapons =
+{
+	"WeaponLobSpecial", "WeaponStaffBall", "WeaponDaggerThrow", "WeaponTorchSpecial", "WeaponAxeSpecial", "WeaponSuitRanged"
+}
+
+game.WeaponData.WeaponBlink.OnFiredFunctionNames = game.WeaponData.WeaponBlink.OnFiredFunctionNames or {}
+
+table.insert(game.WeaponData.WeaponBlink.OnFiredFunctionNames, _PLUGIN.guid .. "." .. "RecordBlinkCharge")
+
+function mod.RecordBlinkCharge(unit, weaponData, args, triggerArgs)
+	if game.CurrentRun.Hero.Weapons["WeaponLob"] and not game.CurrentRun.Hero.Weapons["WeaponLobSpecial"] then
+		local ammoPacks  = game.GetIdsByType({ Name = "LobAmmoPack"})
+		game.SetObstacleProperty({ Property = "Magnetism", Value = game.WeaponData.WeaponLobSpecial.MagnetismMultiplier, DestinationIds = ammoPacks, ValueChangeType = "Multiply" })
+		game.SessionMapState.MagnetismMultiplier = game.WeaponData.WeaponLobSpecial.MagnetismMultiplier
+	end
+end
+
+game.OnBlinkFinished{ "WeaponBlink", function (triggerArgs)
+	if game.CurrentRun.Hero.Weapons["WeaponLob"] and not game.CurrentRun.Hero.Weapons["WeaponLobSpecial"] and game.SessionMapState.MagnetismMultiplier then
+		for id, data in pairs( game.SessionMapState.AutoMagnetizeIds ) do
+			data.MagnetismMultiplier = game.WeaponData.WeaponLobSpecial.MagnetismMultiplier
+		end
+
+		local playerMagnetism = game.SessionMapState.MagnetismMultiplier * game.GetBaseDataValue({ Type = "Obstacle", Name = "LobAmmoPack", Property = "Magnetism"})
+		local ammoPacks  = game.GetIdsByType({ Name = "LobAmmoPack"})
+		for _, ammoId in pairs( ammoPacks ) do
+			if game.GetDistance ({ Id = ammoId, DestinationId = game.CurrentRun.Hero.ObjectId }) >= playerMagnetism then
+				game.SetObstacleProperty({ Property = "Magnetism", Value = 1/game.SessionMapState.MagnetismMultiplier, DestinationId = ammoId, ValueChangeType = "Multiply" })
+			end
+		end
+		game.SessionMapState.MagnetismMultiplier = nil
+	end
+end}
+
+modutil.mod.Path.Wrap("HandleGunBehavior", function (base, weaponData, functionArgs, triggerArgs)
+	base(weaponData, functionArgs, triggerArgs)
+	if game.IsExWeapon( weaponData.Name, { Combat = true }, triggerArgs )  then
+		local chargeStages = game.GetWeaponChargeStages( weaponData )
+		local weaponCharge = (game.MapState.WeaponCharge or {})[weaponData.Name] or 1
+		if #chargeStages <= weaponCharge and game.Contains( {"WeaponStaffBall", "WeaponDaggerThrow", "WeaponTorchSpecial", "WeaponAxeSpecial", "WeaponAxeSpecialSwing", "WeaponSuitRanged"}, weaponData.Name ) then
+			local dataProperties = game.MergeTables(game.EffectData[functionArgs.EffectName].DataProperties, functionArgs.EffectData)
+			dataProperties.Duration = dataProperties.Duration + game.GetTotalHeroTraitValue("OverheatDurationIncrease")
+			game.ApplyEffect({ DestinationId = game.CurrentRun.Hero.ObjectId, Id = game.CurrentRun.Hero.ObjectId, EffectName = functionArgs.EffectName, DataProperties = dataProperties })
+		end
+	end
+end)